Improve Admin UI and Protect last Admin

+++ app/api/admins/[id]/route.ts
@@ -17,7 +17,20 @@ export async function DELETE(req: Request, context: Context) {
   const { id } = await context.params;
 
   await connectDB();
+  const adminCount = await User.countDocuments({ role: "admin" });
 
+  const user = await User.findById(id);
+
+  if (!user) {
+    return NextResponse.json({ message: "User not found" }, { status: 404 });
+  }
+
+  if (user.role === "admin" && adminCount <= 1) {
+    return NextResponse.json(
+      { message: "You cannot delete the last admin account" },
+      { status: 400 }
+    );
+  }
   await User.findByIdAndDelete(id);
 
   return NextResponse.json({ success: true });
